Move lone remaining function and delete file
